Allow logging in using SOAP without echoing password
Some environments do not have access to a browser with JavaScript
enabled (such as SSH server with w3m). For these environments, one would
need to use a SOAP login. There should be a way to log in without having
to write the password in the command line history.

This patch uses a cross platform package to prevent echoing the password
if a user provides a username.