Specify Test Level and Tests To Run During Deploy
Add support for specifying the test level and tests to run to the push
and import commands.

By default, the test level is NoTestRun.  If any test classes are
specified, the test level is set to RunSpecifiedTests.

The -runalltests option is preserved, and is equivalent to the new
RunAllTestsInOrg test level.