Make src a valid source directory
This was previously merged but reverted. This re-enables the feature
again. It was reverted because there were other locations where the path
was hard coded to look at a directory named 'metadata'. If those
situations arise again, instead of hard coding, they should use
`GetSrcDir()` to obtain the true path. This will not only provide
flexibility for src vs metadata, but allow any future conventions to be
easily supported.

Closes #176