Cleanup of some vars and the basic push command
Nothing that should change behavior. Mostly just for readability.

The change to force.go is pretty unrelated and could be moved into it's
own diff. Overall, I'm not sure why this is here because it makes it so
one can't use the force cli they build since it has a non-working uri