GetSourceDir() returns a less-surprising source directory
GetSourceDir() used to get confused by my ~/src/salesforce directory,
because "src" was a part of the directory name. Now, GetSourceDir()
prefers looking at subdirectories and then walks up the parent branch,
until it finds what it's looking for.

It also relies on the existence of package.xml, instead of just the
name of the directory itself.

Finally, GetSourceDir() no longer searches for "src" exhaustively
because looking for "metadata".