Run Anonymous Apex in Test Context
Add -test option to the apex command to run anonymous apex in a test
context by wrapping it in an apex class.  The compileAndTest SOAP call
is used to perform a check-only deploy of the generated class and run
the apex passed in within a single test method.  The test class has
seeAllData=true so it works similarly to anonymous apex.