Easy logout when only one username is logged in.
Change the behaviour of the “logout” command to make the username
optional when only one user is logged in.

This does not change the behaviour of the command when a username is
given. Notably, invocation of the command with no username specified
may or may not exit with error; but the command never exits with error
when a username is specified.