Stream CSV and JSON Query Results
Write csv- and json-formatted query results to stdout as they are
received rather than waiting for all results to be retrieved.

Fixes out-of-memory errors when querying large result sets.

Note that all results are still retrieved when writing to stdout in the
default format because the column widths need to be calculated from the
results.

Update Force.QueryAndSend to accept optional QueryOptions parameters to
support queryAll and tooling options.