Report dangling arguments
Added `MaxExpectedArgs` to `command.Command`. After parsing flags, if the remaining arguments are greater than expect, exit with an error status.

This will prevent situations such as `force fetch -t Metadata MetadataName` where the default action is not probably not the expected outcome.

For commands that already check their arguments or take a variadic amount of arguments `MaxExpectedArgs` is -1.