Add http callback support, cleanup (LP 56260286)
The original goal of this PR was to add support for a callback,
passed into RetrieveBulkJobQueryResultsWithCallback,
that would be called with `*http.Response`,
so the caller could get at the raw response body.

However, doing this required either copying and modifying some code
(making the code worse), or doing a yak-shave to modify most
of the HTTP handling. Of course I went with the yak shave.

This rewrites all of the HTTP GET callers into a single method
which parameters are passed into. In theory the other HTTP verb
methods should also use this; but I didn't need to touch them,
so I left them for now.

So conceptually, instead of `httpGetJSON` and `httpGetJSONAndSend`
and all variations of content types and other specializations,
we just have something akin to
`newAuthedHttpInput(...) httpRequestInput` and
`makeHttpRequest(httpRequestInput)`. The callers are responsible
for configuring the http request input options appropriately.

Additionally, we have some different uses of the term
"content type." Specifically, for bulk work, it's the strings
"JSON", "CSV", and "XML", but elsewhere it's "application/json",
"text/csv", and "application/xml".
We introduce JobContentType and ContentType types to represent these,
and try to avoid using the string content type
(which is still on JobInfo since I didn't want to break
backwards compat).

Over the course of this refactor, I found a couple bugs:

- We were using `X-Sfdc-Session` as "Bearer: <token>" in some places,
  and as simply "<token>" in others.
- Rate limiting was only handled on some code paths, and not others.
- I am pretty sure some of the post/patch/etc methods call the
  wrong method for their retry; but things so happen to work anyway.
- One of the json bind variables was wrong

Because this was an entirely internal refactor,
I only added a test for the new method that takes the http callback.
This is also our first ghttp test in the repo,
so more should follow.

This one method fortunately covers a large part of the changes;
most of the rest are covered by the static typing.
But it warrants some further testing and a close review anyway.