Add an extensible HTTP Request/Response system
In https://github.com/ForceCLI/force/pull/598,
support for an `HttpCallback` was added to some methods,
which gave callers access to the raw `*http.Response`
for things like streaming in the body.

This was only added for POST/PUT/PATCH,
because it used separate codepaths to GET requests.

This takes things a step further,
and adds a configurable `lib.Request` and `lib.Response` type
that can be used by callers.
See request.go for more details.