Overhaul http handling, especially bulk
There is a bug in most of the code that manually unmarshals
an HTTP response into a LoginFault. Mainly, it assumes that,
if the response has 'no results' (empty list, or missing Id),
it is an error; so it unmarshals the LoginFault and uses that
as the error.

This is wrong on multiple levels:

- Actual LoginFaults are already detected at a lower level,
  during the HTTP calls. It is impossible that most of the code
  doing manual LoginFault handling was ever hit in the case
  of a LoginFault. (I suspect the code was at one point not
  provably wrong like it is now).
- As mentioned, empty lists would trigger the branch causing the
  response to be interpreted as an error.
- But the error message would be ": ", because the LoginFault
  did not parse properly.

Additionally, there were a number of latent bugs regarding ignoring
error returns from marshal/unmarshal calls.
This could result in not serializing fields to pass to the server,
or incorrectly parsing the response and returning empty results.

This commit takes another pass at cleaning up ForceCLI http management.
Namely, it does the following:

- Adds an `internal` module which has marshal/unmarshal wrappers
  the provide a clearer error message. The errors include a snippet
  of the content that failed to marshal/unmarshal to aid diagnostics,
  since in most cases this will be marshaling a bad type,
  or unmarshling the wrong content type.
- Removes the rabbit hole of "httpPostJson", "httpPatchJson", etc.,
  which were all just variations of the same HTTP method
  (calling POST/PATCH with a content type), with most of the lines
  of code in each method being pure copy/paste.
  There is now just an `httpPostPatch` method that takes a content type
  and http method.
- One http handling method already did a more thorough error handling/
  coercion. This code was pulled out and is now available in this
  additional http method. Ideally all http calls would go through the
  same central handler, but we're not there yet
  (and because we are calling different APIs, this is not so easy).