Improve Query Output With Empty Relationship Values
Update the default console table output to only output "(null)" when a
field is null.  Output an empty string when there is output for the row.

This changes the output when querying relationships.

For example, assume the field My_Relationship__r.My_Field__c is queried and
there are two rows in the output.

If My_Relationship__c is null, "(null)" will be displayed for
My_Relationship__r and nothing will be displayed for
My_Relationship__r.My_Field__c.

If My_Relationship__c is not null, nothing will be displayed for
My_Relationship__r (which is only included in the output if there is
another row where My_Relationship__c is null).
My_Relationship__r.My_Field__c will still display "(null)" if
My_Field__c is null.