Add Smart Flow Deploy Option
Add --smart-flow-version flag to push and import commands to enable
automatic selection of the next flow version and deletion of inactive
versions.  This results in there only being two version of the flow in
the org, the newly deployed version and the previously deployed version.

The lowest available version number will be used.  For example, if
versions 2 and 3 are in the org, and 3 is active, the new version will
be 1, and version 2 will be deleted.